feat: 나의 예약 페이지 UI 수정, bootpay메서드 코드 pay_screen 으로 이동
