+++ lib/presentation/mybooks/my_books_screen.dart
@@ -47,137 +47,145 @@ class MyBooksScreen extends StatelessWidget {
                               child: Container(
                                 margin: const EdgeInsets.all(8),
                                 decoration: BoxDecoration(
-                                  color: Colors.grey,
+                                  color: Colors.white,
                                   borderRadius: BorderRadius.circular(12.0),
                                 ),
                                 padding: const EdgeInsets.all(10),
                                 child: Row(
-                                  crossAxisAlignment: CrossAxisAlignment.center,
                                   mainAxisAlignment:
                                       MainAxisAlignment.spaceBetween,
                                   children: [
-                                    Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      mainAxisAlignment:
-                                          MainAxisAlignment.spaceBetween,
-                                      children: [
-                                        Text(
-                                          '예약번호: ${myBookItem.bookId}',
-                                          style: const TextStyle(
-                                              fontSize: 15,
-                                              color: Colors.black),
-                                        ),
-                                        const SizedBox(
-                                          height: 10,
-                                        ),
-                                        Text(
-                                          '탑승일: ${myBookItem.flightDate}',
-                                          style: const TextStyle(
-                                              fontSize: 13,
-                                              color: Colors.black),
-                                        ),
-                                      ],
+                                    Expanded(
+                                      flex: 9,
+                                      child: Column(
+                                        children: [
+                                          Row(
+                                              mainAxisAlignment:
+                                                  MainAxisAlignment
+                                                      .spaceBetween,
+                                              children: [
+                                                Text(
+                                                  '예약번호: ${myBookItem.bookId}',
+                                                  style: const TextStyle(
+                                                      fontSize: 15,
+                                                      color: Colors.black),
+                                                ),
+                                                Row(
+                                                  children: [
+                                                    Text(
+                                                      myBookItem.departureCode!,
+                                                      style: const TextStyle(
+                                                          color: Colors.black,
+                                                          fontWeight:
+                                                              FontWeight.bold),
+                                                    ),
+                                                    const Padding(
+                                                      padding: EdgeInsets.only(
+                                                          left: 8.0),
+                                                      child: Icon(
+                                                        Icons.flight_takeoff,
+                                                        color: Colors.pink,
+                                                      ),
+                                                    ),
+                                                    Padding(
+                                                      padding:
+                                                          const EdgeInsets.only(
+                                                              left: 8.0),
+                                                      child: Text(
+                                                        myBookItem.arrivalCode!,
+                                                        style: const TextStyle(
+                                                            color: Colors.black,
+                                                            fontWeight:
+                                                                FontWeight
+                                                                    .bold),
+                                                      ),
+                                                    )
+                                                  ],
+                                                ),
+                                              ]),
+                                          const SizedBox(
+                                            height: 10,
+                                          ),
+                                          Row(
+                                              crossAxisAlignment:
+                                                  CrossAxisAlignment.center,
+                                              mainAxisAlignment:
+                                                  MainAxisAlignment
+                                                      .spaceBetween,
+                                              children: [
+                                                Text(
+                                                  '탑승일: ${myBookItem.flightDate}',
+                                                  style: const TextStyle(
+                                                      fontSize: 13,
+                                                      color: Colors.black),
+                                                ),
+                                                Text(
+                                                  '탑승자명: ${myBookItem.passengerName}',
+                                                  style: const TextStyle(
+                                                      fontSize: 13,
+                                                      color: Colors.black),
+                                                ),
+                                              ]),
+                                        ],
+                                      ),
                                     ),
-                                    Column(
-                                      crossAxisAlignment:
-                                          CrossAxisAlignment.start,
-                                      children: [
-                                        Row(
-                                          children: [
-                                            Text(
-                                              myBookItem.departureCode!,
-                                              style: const TextStyle(
-                                                  color: Colors.black,
-                                                  fontWeight: FontWeight.bold),
-                                            ),
-                                            const Padding(
-                                              padding:
-                                                  EdgeInsets.only(left: 8.0),
-                                              child: Icon(
-                                                Icons.flight_takeoff,
-                                                color: Colors.pink,
-                                              ),
-                                            ),
-                                            Padding(
-                                              padding: const EdgeInsets.only(
-                                                  left: 8.0),
-                                              child: Text(
-                                                myBookItem.arrivalCode!,
-                                                style: const TextStyle(
-                                                    color: Colors.black,
-                                                    fontWeight:
-                                                        FontWeight.bold),
-                                              ),
-                                            )
-                                          ],
-                                        ),
-                                        const SizedBox(
-                                          height: 10,
-                                        ),
-                                        Text(
-                                          '탑승자명: ${myBookItem.passengerName}',
-                                          style: const TextStyle(
-                                              fontSize: 13,
-                                              color: Colors.black),
-                                        ),
-                                      ],
-                                    ),
-                                    (myBookItem.payStatus! == 1)
-                                        ? const Column(
-                                            mainAxisAlignment:
-                                                MainAxisAlignment.spaceBetween,
-                                            children: [
-                                              Text('결제',
+                                    SizedBox(
+                                        width:
+                                            MediaQuery.of(context).size.width *
+                                                0.15),
+                                    Expanded(
+                                      flex: 1,
+                                      child: Column(
+                                        mainAxisAlignment:
+                                            MainAxisAlignment.spaceBetween,
+                                        children: [
+                                          Text('결제',
+                                              style: TextStyle(
+                                                  color:
+                                                      (myBookItem.payStatus! ==
+                                                              1)
+                                                          ? Colors.grey
+                                                          : Colors.black)),
+                                          const SizedBox(
+                                            height: 10,
+                                          ),
+                                          (myBookItem.payStatus! == 1)
+                                              ? const Text('완료',
                                                   style: TextStyle(
-                                                      color: Colors.black)),
-                                              SizedBox(
-                                                height: 10,
-                                              ),
-                                              Text('완료',
-                                                  style: TextStyle(
-                                                      color: Colors.black)),
-                                            ],
-                                          )
-                                        : Column(
-                                            crossAxisAlignment:
-                                                CrossAxisAlignment.center,
-                                            children: [
-                                              const Text('결제',
-                                                  style: TextStyle(
-                                                      color: Colors.black)),
-                                              const SizedBox(
-                                                height: 10,
-                                              ),
-                                              RoundCheckBox(
-                                                checkedWidget: const Icon(
-                                                    Icons.check,
-                                                    size: 13),
-                                                checkedColor:
-                                                    const Color(0xFF9bc6bf),
-                                                size: 14,
-                                                isChecked: (viewModel
-                                                        .forPaymentList
-                                                        .contains(myBookItem))
-                                                    ? true
-                                                    : false,
-                                                onTap: (selected) {
-                                                  viewModel
-                                                      .forPaymentCheckBoxTap(myBookItem);
-                                                },
-                                              )
-                                            ],
-                                          )
+                                                      color: Colors.grey))
+                                              : RoundCheckBox(
+                                                  isRound: false,
+                                                  checkedWidget: const Icon(
+                                                      Icons.check,
+                                                      size: 14),
+                                                  uncheckedColor: Colors.grey,
+                                                  checkedColor: Colors.green,
+                                                  size: 14,
+                                                  isChecked: (viewModel
+                                                          .forPaymentList
+                                                          .contains(myBookItem))
+                                                      ? true
+                                                      : false,
+                                                  onTap: (selected) {
+                                                    viewModel
+                                                        .forPaymentCheckBoxTap(
+                                                            myBookItem);
+                                                  },
+                                                )
+                                        ],
+                                      ),
+                                    )
                                   ],
                                 ),
                               ),
                             );
-                          },
-                        ),
+                          }),
                 ),
                 ElevatedButton(
                   onPressed: () {
-                    context.go('/pay', extra: viewModel.forPaymentList);
+                    if (viewModel.forPaymentList.isNotEmpty) {
+                      context.go('/pay', extra: viewModel.forPaymentList);
+                    }
                   },
                   style: ButtonStyle(
                     //테두리 모양 조절
